# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Codeplane API
  version: 1.0.0
  description: API for Codeplane - a jj-native development platform
servers:
  - url: https://api.codeplane.app
paths:
...
  /repos/{owner}/{repo}:
    get:
      summary: GET /repos/{owner}/{repo}
      operationId: get_repos__owner___repo_
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    patch:
      summary: PATCH /repos/{owner}/{repo}
      operationId: patch_repos__owner___repo_
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: DELETE /repos/{owner}/{repo}
      operationId: delete_repos__owner___repo_
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...
  /orgs/{org}/repos:
    post:
      summary: POST /orgs/{org}/repos
      operationId: post_orgs__org__repos
      tags:
        - Orgs
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    get:
      summary: GET /orgs/{org}/repos
      operationId: get_orgs__org__repos
      tags:
        - Orgs
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...
  /repos/{owner}/{repo}/git/trees/{sha}:
    get:
      summary: GET /repos/{owner}/{repo}/git/trees/{sha}
      operationId: get_repos__owner___repo__git_trees__sha_
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: sha
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
  /repos/{owner}/{repo}/git/commits/{sha}:
    get:
      summary: GET /repos/{owner}/{repo}/git/commits/{sha}
      operationId: get_repos__owner___repo__git_commits__sha_
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: sha
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
  /repos/{owner}/{repo}/git/refs:
    get:
      summary: GET /repos/{owner}/{repo}/git/refs
      operationId: get_repos__owner___repo__git_refs
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
  /repos/{owner}/{repo}/topics:
    get:
      summary: GET /repos/{owner}/{repo}/topics
      operationId: get_repos__owner___repo__topics
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    put:
      summary: PUT /repos/{owner}/{repo}/topics
      operationId: put_repos__owner___repo__topics
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
  /repos/{owner}/{repo}/stargazers:
    get:
      summary: GET /repos/{owner}/{repo}/stargazers
      operationId: get_repos__owner___repo__stargazers
      tags:
        - Repos
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
  /user/starred/{owner}/{repo}:
    put:
      summary: PUT /user/starred/{owner}/{repo}
      operationId: put_user_starred__owner___repo_
      tags:
        - User
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: DELETE /user/starred/{owner}/{repo}
      operationId: delete_user_starred__owner___repo_
      tags:
        - User
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...
  /orgs/{org}/teams/{team}/repos/{owner}/{repo}:
    put:
      summary: PUT /orgs/{org}/teams/{team}/repos/{owner}/{repo}
      operationId: put_orgs__org__teams__team__repos__owner___repo_
      tags:
        - Orgs
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: team
          in: path
          required: true
          schema:
            type: string
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: DELETE /orgs/{org}/teams/{team}/repos/{owner}/{repo}
      operationId: delete_orgs__org__teams__team__repos__owner___repo_
      tags:
        - Orgs
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: team
          in: path
          required: true
          schema:
            type: string
        - name: owner
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
...